+++ rapdev_sophos/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - rapdev_sophos
 
+## 1.1.0 / 2022-05-06
+* [Added] Added capabilities for bringing in SIEM events and SIEM logs
+* [Added] Added dd_service for both common and SIEM alerts/event logs
+
 ## 1.0.1 / 2022-02-03
 * [Bugfix] Fix bug causing integration not to start